Added encoder section of improved unet
Implemented the encoder portion of the improved unet based on the description/graph of the network from the research paper. Noticed that the output filters for each layer of the encoder doubled in size and so a variable 'n_filter' gets doubled for each successive layer. Used the VAE implemented from demo 2 as a base and reference.