Added Modules & Train, Altered Dataset
Added in the Modules file containing the current and base implementation of the ViT based model in addition to the training file which implements this model and training there of.

The Dataset file was altered to solely have a singular train test split as apposed to train val test split.