Massive overhaul of unet structure
Redid the entire implementation since I couldn't get it working and the messy structure made it hard to debug the forward pass of the unet. Removed redundant upscale layers and only implemented 1 to be reused since the unet only requires the same upscaling factor. Changed the batch normalisations to instance normalisations and the relu to leaky relu since that was what was specified in the paper https://arxiv.org/pdf/1802.10508v1.pdf. Also added the segmentations layers and the softmax layer for the output. Also changed dropout from 3d to 2d and did same as upscale layer and only made one layer and reused for the rest of the network. Implemented the forward pass as well with everything that was described in the paper (element wise sums, no convolution transposes or max pooling).