Modified to only load images/masks with "ISIC" prefix
Modified the custom dataset to only load images/masks with "ISIC" prefix as it was noticed that the dataset on rangapur had kept the attribution text files which flagged errors when trying to run code.
Also modified the class so that the images and masks have their own transforms. This is mainly so that the images can be normalised through a transform without affecting the masks.