Changed class back to only use one transform for the images and masks
Originally had separate transforms for the images and masks since it was expected that the images would require normalisation while the masks would not however that does not seem to be the case. The image RGB values are already within the range of 0 to 1 and thus the images and masks can use the same transform. This also makes it more convenient for any data augmentation to be applied.