Critical update: changed softmax layer to sigmoid layer
This makes a major change to the model since it replaces the output layer of softmax to a sigmoid. Softmax would work with multiclass segmentation however the project task is binary class segmentation. Previously the model failed to learn and spits out blank mask segmentation predictions but changing the output layer to the sigmoid allows the model to learn and give reasonable outputs since there is a single output channel.