Consolidated image and mask transforms, added dataloader for validation set, added validation loop to training section
Combined the image and mask transforms together since the images don't require normalisation as the RGB values are already in the range of 0 to 1. Added the validation loop to the training section which will break out of the loop early once the dice coefficient threshold is reached. Added the transform and the data loader for the validation set, added a list for collecting the validation loss and an extra plot function for adding the validation loss to the training loss plot.