Added comments and header block for explaining code
Added in more comments to explain parts of the code and for separating the different sections of the code that do a different function (e.g. data loading section, model training section). Also added a header block for explaining what the file is about and does as well as some notes to keep in mind when using the file.