Added header block and fixed mistake in comments
Added header block describing what the file is used for and some notes to keep in mind when using the file. Also fixed a mistake in the comments that had height and width the other way around when describing the structure of tensors (e.g. had it as [n, c, w, h] instead of [n, c, h, w]).